drivers: ad400x: Fix buffer mask for 18 and 20 bits
For the chips that support 18 or 20 bit resolution,
the least significant bits are in the higher nibble
of the last byte. Sign needs to be taken into consideration
when printing the result.

While here, do the bit operations in the project as this
is where the print happens. By doing so, we can avoid
unneded bit operations in the driver when iio is added
lter, and we can use scan_type to porperly descibe the
part bit format.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>